requirements.txt
mysql.connector
PyPDF2
streamlit